Add files via upload
This is an additional .ipynb file which will work in the backend (not integrated to the frontend as of now ). This uses DocQuery for document query processing and gives best results but it is not integrated to the frontend. The frontend team can use this for better result as compared to the regex one provided in app.py and integrated with the html files in template folder.